timemachine: Manual change - add closing quote mark to chapter-final extended quotes
To ensure that the CHAPTER. title is *not* considered part of the quote and to avoid problems with regions_invert() not properly handling the overlap of quotes and chapter.text at the beginning of the next chapter. Future CLiC can consider whether the definition/implementation of extended quotes should allow them to cross chapter boundaries.